Ignore --device option in android build
iOS uses the `--device` flag to indicate that it should build for the
device rather than the simulator. Until recently, Android simply ignored
this option. Now it throws an error.

This causes problems for people targeting multiple platforms with the CLI
tools, because a command like `cordova build --release --device` will now
fail on Android, rather than building both Android and iOS versions of the
app.