solve issue where onActivityResult gets called before onResume
if one of the plugins launches startActivityForResult and the Android OS decides to kill our MainActivity. what happens when the launched activity is closed.
the previous activity (our MainActivity is relaunched and onActivityResult is called, our mainActivity however has not yet had the time to load all onload=true plugins so therefore the savedInstanceState callBackService can not be called.