cdvBuildMultipleApks casting fixed
Casting `cdvBuildMultipleApks` via `!!` does not work when you want to modify the build process via the command line. The problem can be simply reproduced when running: `cordova build android -- --gradleArg=-PcdvBuildMultipleApks=false`. The expected output would be a build with a single apk. But the result is a build with multiple apks cause `!!"false"` is `true`. This fix solves the problem by using an more mature casting process that also can handle string that contain the value `"false"`.