[CB-10600] 'cordova run --release' uses wrong apk
When doing `cordova run android --release`, the CLI attempts to use the
android-release-unsigned.apk, which naturally fails to install on the
device because it is unsigned. The release build does also generate the
properly signed android-release.apk, and that should be used instead.

The apk-finding logic for "debug" excludes unsigned builds, so the
"release" logic should follow a similar pattern.

Fixes https://issues.apache.org/jira/browse/CB-10600