CB-8383 Handle pause event when keeprunning=false
I added this pull request to hopefully get some attention to this bug https://issues.apache.org/jira/browse/CB-8383. When keeprunning = false in config.xml, any pause event processing is halted until the app returns. Not only is the pause event missed on app backgrounding, but on return it processes the pause and resume events.